requirements.txt
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-framework==0.1.0
django-simple-history==2.11.0
djangorestframework==3.11.0
gunicorn==20.0.4
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.4
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
whitenoise==5.1.0
